+++ tests/test_examples.py
@@ -1,77 +1,139 @@
-from typing import Dict, Set
-
-import os
-import subprocess
-import re
-import pytest
-import json
 import itertools
-from pylint import lint
-from io import StringIO
+import json
+import os
+import re
 import sys
+from io import StringIO
+from typing import Dict, List, Set, Union
 
+import pytest
+from pylint import lint
+
+import python_ta
 
 _EXAMPLES_PATH = "examples/pylint/"
-_EXAMPLE_PREFIX_REGEX = r"[CEFRW]\d{4}"
+_CUSTOM_CHECKER_PATH = "examples/custom_checkers/"
+_PYCODESTYLE_PATH = "examples/custom_checkers/e9989_pycodestyle/"
+
+_EXAMPLE_PREFIX_REGEX = r"[cerfw]\d{4}"
+_PYCODESTYLE_PREFIX_REGEX = r"^e\d{3}_(error|no_error)\.py$"
 
 
 # The following tests appear to always fail (further investigation needed).
 IGNORED_TESTS = [
     "e1131_unsupported_binary_operation.py",
     "e0118_used_prior_global_declaration.py",
-    "w0125_using_constant_test.py",
     "w0631_undefined_loop_variable.py",
     "w1503_redundant_unittest_assert.py",
     "e1140_unhashable_dict_key.py",
     "r0401_cyclic_import.py",  # R0401 required an additional unit test but should be kept here.
+    "e9999_forbidden_import_local.py",  # This file itself (as an empty file) should not be tested
+    "c9104_ModuleNameViolation.py",  # Due to different naming format, this file is handled separately
+    "e0643_potential_index_error.py",
+    "e1003_bad_super_call.py",
+    "e1143_unhashable_member.py",
+    "r0201_no_self_use.py",
+    "e9950_forbidden_python_syntax.py",
 ]
 
 
-def get_file_paths():
-    """Gets all the files from the examples folder for testing. This will
-    return all the full file paths to the file, meaning they will have the
-    _EXAMPLES_PATH prefix followed by the file name for each element.
-    A list of all the file paths will be returned."""
+def get_file_paths(paths: Union[str, List[str]]) -> List[str]:
+    """
+    Get all the Python files from the specified directories for testing. This will
+    return the full file paths for each Python file, excluding those listed in IGNORED_TESTS.
+    The file paths will have the directory path prefixed to the file name for each element.
+    A list of all the file paths will be returned.
+
+    :param paths: The path or list of paths to retrieve the Python files from.
+    :return: A list of full file paths to the Python files.
+    """
     test_files = []
-    for _, _, files in os.walk(_EXAMPLES_PATH, topdown=True):
-        for filename in files:
-            if filename not in IGNORED_TESTS and filename.endswith(".py"):
-                test_files.append(_EXAMPLES_PATH + filename)
+
+    if isinstance(paths, str):
+        paths = [paths]
+
+    for path in paths:
+        for root, _, files in os.walk(path, topdown=True):
+            for filename in files:
+                if filename not in IGNORED_TESTS and filename.endswith(".py"):
+                    full_path = os.path.join(root, filename)
+                    rel_path = os.path.relpath(full_path, path)
+                    test_files.append(os.path.join(path, rel_path))
+
     return test_files
 
 
-@pytest.fixture(scope="session", autouse=True)
-def symbols_by_file() -> Dict[str, Set[str]]:
-    """Run pylint on all the example files and return the map of file name to the
-    set of Pylint messages it raises."""
+def _symbols_by_file_pyta(paths: List[str], include_msg: bool = False) -> Dict[str, Set[str]]:
+    """
+    Run python_ta.check_all() on files from specified directories and return the map of file name to the
+    set of PythonTA messages it raises. If include_msg is set True, PythonTA message descriptions are
+    included along with message symbols.
 
+    :param paths: The paths to retrieve the files from.
+    :param include_msg: whether to include message descriptions in the symbol set
+    :return: A dictionary mapping each file name to a set of PythonTA message symbols
+    (and descriptions if include_msg is True).
+    """
     sys.stdout = StringIO()
-    lint.Run(
-        [
-            "--reports=n",
-            "--rcfile=python_ta/config/.pylintrc",
-            "--output-format=json",
-            *get_file_paths()
-        ], exit=False
+    python_ta.check_all(
+        module_name=get_file_paths(paths),
+        config={
+            "output-format": "python_ta.reporters.JSONReporter",
+        },
     )
+
     jsons_output = sys.stdout.getvalue()
     sys.stdout = sys.__stdout__
-
-    pylint_list_output = json.loads(jsons_output)
+    pyta_list_output = json.loads(jsons_output)
 
     file_to_symbol = {}
-    for path, group in itertools.groupby(pylint_list_output, key=lambda d: d["path"]):
-        symbols = {message["symbol"] for message in group}
-        file = os.path.basename(path)
+    for path, group in itertools.groupby(
+        pyta_list_output, key=lambda d: os.path.basename(d["filename"])
+    ):
+        symbols = set()
+        for message in group:
+            for msg in message["msgs"]:
+                symbols.add(msg["symbol"])
+                if include_msg:
+                    symbols.add(msg["msg"])
 
+        file = os.path.basename(path)
         file_to_symbol[file] = symbols
 
     return file_to_symbol
 
 
-@pytest.mark.parametrize("test_file", get_file_paths())
-def test_examples_files(test_file: str, symbols_by_file: Dict[str, Set[str]]) -> None:
-    """Creates all the new unit tests dynamically from the testing directory."""
+@pytest.fixture(scope="session")
+def pyta_examples_symbols() -> Dict[str, Set[str]]:
+    """
+    A pytest fixture that runs once per test session.
+    This fixture analyzes example files using python_ta and returns a dictionary mapping each file name
+    to the set of PythonTA message symbols raised.
+
+    :return: A dictionary mapping file names to sets of PythonTA message symbols.
+    """
+    return _symbols_by_file_pyta([_EXAMPLES_PATH, _CUSTOM_CHECKER_PATH])
+
+
+@pytest.fixture(scope="session")
+def pyta_pycodestyle_symbols() -> Dict[str, Set[str]]:
+    """
+    A pytest fixture that runs once per test session.
+    This fixture analyzes pycodestyle error test cases using python_ta and returns a dictionary mapping each file name
+    to the set of PythonTA message symbols and descriptions.
+
+    :return: A dictionary mapping file names to sets of PythonTA message symbols and descriptions.
+    """
+    return _symbols_by_file_pyta([_PYCODESTYLE_PATH], include_msg=True)
+
+
+@pytest.mark.parametrize("test_file", get_file_paths([_EXAMPLES_PATH, _CUSTOM_CHECKER_PATH]))
+def test_examples_files_pyta(test_file: str, pyta_examples_symbols: Dict[str, Set[str]]) -> None:
+    """
+    Dynamically creates and runs unit tests for Python files in the examples and custom checker directories.
+    This test function deduces the error type from the file name and checks if the expected error message is present
+    in PythonTA's report.
+    """
     base_name = os.path.basename(test_file)
     if not re.match(_EXAMPLE_PREFIX_REGEX, base_name[:5]):
         return
@@ -80,10 +142,72 @@ def test_examples_files(test_file: str, symbols_by_file: Dict[str, Set[str]]) ->
     checker_name = base_name[6:-3].replace("_", "-")  # Take off prefix and file extension.
 
     test_file_name = os.path.basename(test_file)
-    file_symbols = symbols_by_file[test_file_name]
+    file_symbols = pyta_examples_symbols[test_file_name]
 
     found_pylint_message = checker_name in file_symbols
-    assert found_pylint_message, f"Failed {test_file}. File does not add expected message."
+    assert (
+        found_pylint_message
+    ), f"Failed {test_file}. File does not add expected message  {file_symbols}."
+
+
+@pytest.mark.parametrize("test_file", get_file_paths(_PYCODESTYLE_PATH))
+def test_pycodestyle_errors_pyta(
+    test_file: str, pyta_pycodestyle_symbols: Dict[str, Set[str]]
+) -> None:
+    """
+    Dynamically creates and runs unit tests for pycodestyle error test cases.
+    This test function deduces the PEP8 error code from the file names. It checks if pycodestyle error is present
+    in PythonTA's report and if the correct PEP8 error type is in the message description.
+    """
+    base_name = os.path.basename(test_file)
+    if not re.match(_PYCODESTYLE_PREFIX_REGEX, base_name.lower()):
+        return
+    if not base_name.lower().endswith(".py"):
+        assert False
+
+    # skip the test case if it does not have errors
+    has_error = base_name[5:] == "error.py"
+    if not has_error:
+        return
+
+    error_code = base_name[:4].upper()  # get the specific PEP8 error code
+    test_file_name = os.path.basename(test_file)
+    file_symbols = pyta_pycodestyle_symbols[test_file_name]
+
+    found_pycodestyle_message = "pep8-errors" in file_symbols
+    assert found_pycodestyle_message, f"Failed {test_file}. File does not add expected message."
+    assert any(
+        error_code in msg for msg in file_symbols
+    ), f"Failed {test_file}. The correct PEP8 error type is not in reported message."
+
+
+def test_c9104_module_name_violation() -> None:
+    """
+    Test that examples/custom_checkers/c9104_ModuleNameViolation.py adds C9104 module-name-violation.
+    This test is separate as the naming convention for this file is different from the rest of the examples.
+    """
+    module_name_violation = "examples/custom_checkers/c9104_ModuleNameViolation.py"
+    sys.stdout = StringIO()
+    python_ta.check_all(
+        module_name=module_name_violation,
+        config={
+            "output-format": "python_ta.reporters.JSONReporter",
+        },
+    )
+
+    jsons_output = sys.stdout.getvalue()
+    sys.stdout = sys.__stdout__
+    pyta_list_output = json.loads(jsons_output)
+
+    message_symbols = []
+    for message in pyta_list_output:
+        for msg in message["msgs"]:
+            message_symbols.append(msg["symbol"])
+
+    found_module_name_violation = "module-name-violation" in message_symbols
+    assert (
+        found_module_name_violation
+    ), f"Failed {module_name_violation}. File does not add expected message."
 
 
 def test_cyclic_import() -> None:
@@ -106,8 +230,10 @@ def test_cyclic_import() -> None:
             "--reports=n",
             "--rcfile=python_ta/config/.pylintrc",
             "--output-format=json",
-            cyclic_import_helper, cyclic_import_file
-        ], exit=False
+            cyclic_import_helper,
+            cyclic_import_file,
+        ],
+        exit=False,
     )
     jsons_output = sys.stdout.getvalue()
     sys.stdout = sys.__stdout__